Rearchitechtured the vm, slots, gc and new ffi
Slots are nothing but an index of the datastack, and from now, all
variables and values are stored and accessed from the datastack
directly, insted of mudding over several `Env` and `Record` structs. The
parser now keeps track for required slots for local variables of each
functions and emits the instruction to reserve them as the first
instruction of the function. Even the CALL instruction now pushes the
return address in the stack, along with the `baseptr`, which is the
index of the first slot of the previous function, from which all local
slots are referenced, and sets the `baseptr` of the present function to
be `oldbaseptr + 1`.

Likewise, `Instance` and `Array` representations are also updated to
keep with the new slot based architechture and also to support garbage
collection. Each pointer of `Instance`, `String` and `Array` now holds
an `Object` as its first member. This `Object` is responsible for the
refcounted automatic garbage collection, controlled by
GC_THRESHOLD_BYTES. When size of allocated objects exceedes the
GC_THRESHOLD_BYTES, in the next call to `obj_alloc`, `obj_gc` is
triggered, which collects and frees all objects whose refcounts are 0.

The foreign function interface has also been updated to keep up with the
changes, and `foreign_function.h` now provides a fully opaque API for
interacting with Alang and native objects, and provides a much better
and strict abstraction over the existing one.

Even with this plethora of changes, the syntax and semantics of the
original language are kept almost unchanged, which means any existing
Alang program should run just as fine as it was before, maybe even show
some speedups in certain cases.

All these being said though, this new release is not perfect. It still
doesn't provide expected speedups in all cases, even slows down
horribly in some trivial ones, and may crash at times. All of which should be easier to track down
from now on, as the base has been reconstructed from ground up to be
more strong and structurally sound, and as efficient as possible.